compositor: Skip pads that are completely obscured by a higher zorder pad
For each frame, compare the frame boundaries, check opacity, and skip the frame
if it's going to be completely overwritten by a higher zorder frame. The check
is O(n^2), but that doesn't matter here because the number of sinkpads is small.

More can be done to avoid needless drawing, but this covers the majority of
cases. See TODOs. Ideally, a reverse painter's algorithm should be used for
optimal drawing, but memcpy during compositing is small compared to the CPU used
for frame conversion on each pad.